requirements.txt
textarena
httpx
pyyaml
rich
vllm
ollama